fix: prevent integer truncation in HEALPix degrade of observed counts
hist_obs is stored as int32. hp.ud_grade preserves input dtype when
averaging child pixels, so it truncated the per-pixel mean to 0 for
any group whose observed-star sum didn't divide evenly among children
(the common case for sparse counts), before multiplying back up by
n_children. This silently zeroed adaptive-coarsening cells that had
real observations. The Bayesian prior's +1 numerator masked the
symptom in use_prior=True output; with use_prior=False it surfaced as
hard 0% completeness in bins with observed targets.

+++ src/sdssv_apogee_sf/_selfunc.py
@@ -62,7 +62,11 @@ def _restrict_range(hist: np.ndarray, bins: np.ndarray,
 def _degrade_sum(arr1d: np.ndarray, nside_in: int, nside_out: int) -> np.ndarray:
     """Degrade a 1-D HEALPix map from nside_in to nside_out by summing child pixels."""
     n_children = (nside_in // nside_out) ** 2
-    return hp.ud_grade(arr1d, nside_out) * n_children
+    # hp.ud_grade preserves input dtype: for integer maps (e.g. observed-star
+    # counts) it truncates the intermediate per-pixel mean, silently zeroing
+    # out sparse counts that don't divide evenly among the children before
+    # they get multiplied back up. Cast to float first so the mean is exact.
+    return hp.ud_grade(arr1d.astype(float), nside_out) * n_children
 
 
 def _build_adaptive(
